+++ src/CasCap.Common.Extensions/CasCap.Common.Extensions.csproj
@@ -11,4 +11,8 @@
     <PackageTags>extensions,helpers</PackageTags>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CasCap.Common.Logging\CasCap.Common.Logging.csproj" />
+  </ItemGroup>
+
 </Project>
