+++ src/CasCap.Common.Extensions/Extensions/ExtensionHelpers.cs
@@ -513,103 +513,6 @@ public static class ExtensionHelpers
         return string.Empty;
     }
 
-    #region IO extensions
-    public static string Extend(this string root, string folderOrFile)
-    {
-        var path = Path.Combine(root, folderOrFile);
-        return path;
-    }
-
-    //public static string ExtendAndCreateDirectory(this string root, string folderOrFile)
-    //{
-    //    var directory = root.ExtendPath(Path.GetFullPath(folderOrFile));
-    //    if (!Directory.Exists(directory))
-    //        Directory.CreateDirectory(directory);
-    //    return directory;
-    //}
-
-    public static void WriteAllBytes(this string path, byte[] bytes)
-    {
-        var dir = Path.GetDirectoryName(path);
-        if (dir is not null)
-        {
-            if (!Directory.Exists(dir)) Directory.CreateDirectory(dir);
-            File.WriteAllBytes(path, bytes);
-        }
-        else
-            throw new Exception($"GetDirectoryName not possible for path '{path}'");
-    }
-
-    public static void WriteAllText(this string path, string str)
-    {
-        var dir = Path.GetDirectoryName(path);
-        if (dir is not null)
-        {
-            if (!Directory.Exists(dir)) Directory.CreateDirectory(dir);
-            File.WriteAllText(path, str);
-        }
-        else
-            throw new Exception($"GetDirectoryName not possible for path '{path}'");
-    }
-
-    public static void AppendTextFile(this string path, string content)
-    {
-        var dir = Path.GetDirectoryName(path);
-        if (dir is not null)
-        {
-            if (!Directory.Exists(dir)) Directory.CreateDirectory(dir);
-            if (!File.Exists(path))
-            {
-                using var sw = File.CreateText(path);
-                sw.WriteLine(content);
-            }
-            else
-            {
-                try
-                {
-                    using var sw = File.AppendText(path);
-                    sw.WriteLine(content);
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine(ex.ToString());
-                }
-            }
-        }
-        else
-            throw new Exception($"GetDirectoryName not possible for path '{path}'");
-    }
-
-    public static List<string> ReadTextFile(this string path)
-    {
-        var output = new List<string>(50000);
-        if (File.Exists(path))
-        {
-            //var count = TotalLines(path);
-            //output = new List<string>(count);
-            foreach (var line in File.ReadLines(path))
-            {
-                if (string.IsNullOrWhiteSpace(line)) continue;
-                output.Add(line);
-            }
-            //using (var fs = File.Open(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
-            //{
-            //    using (var stream = new StreamReader(fs))
-            //    {
-            //        while (true)
-            //        {
-            //            var line = stream.ReadLine();
-            //            output.Add(line);
-            //            if (line is null)
-            //                break;
-            //        }
-            //    }
-            //}
-        }
-        return output;
-    }
-    #endregion
-
     public static string Clean(this string thisString, string replacement = "")
     {
         return rgx.Replace(thisString, replacement);
